+++ src/components/auth/components/Recaptcha/RecaptchaWeb.js
